IWorkspaceService: derive from RoslynWorkspace
This is the first step in actually providing support for F# and
other languages. This effectively moves all of the Roslyn bits
we're using to be entirely inside the RoslynWorkspace.

It defines a small API for future language/workspace support.

When WorkbookPageViewModel goes away, the RoslynWorkspace itself
can be cleaned up a lot as well.