Remove navigation delay
This was really weird. The delay was never needed. But the page would fail
to show anything in the browser (although it definitely loaded) unless load
occurred while another window was focused.

When I commented-out more UI bits, this behavior stopped, and now the page
always seems to load and display properly.